# openapi.yaml
openapi: 3.0.3
info:
  title: Family Budget API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: Health response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
  /api/v1/users:
    post:
      summary: Register a new family owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request
        '409':
          description: User already exists
  /api/v1/users/{id}:
    get:
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  family:
                    $ref: '#/components/schemas/Family'
        '404':
          description: Not found
  /api/v1/users/{id}/categories:
    get:
      summary: List categories for user family
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
        '404':
          description: Not found
    post:
      summary: Create a category in the user's family
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Created category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Validation error
        '404':
          description: User not found
  /api/v1/users/{id}/categories/{categoryId}:
    put:
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Validation error
        '404':
          description: Not found
  /api/v1/users/{id}/categories/{categoryId}/archive:
    post:
      summary: Toggle archive state for a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryArchiveRequest'
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Validation error
        '404':
          description: Not found
  /api/v1/users/{id}/accounts:
    get:
      summary: List accounts for user family
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
        '404':
          description: Not found
    post:
      summary: Create an account in the user's family
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Created account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Validation error
        '404':
          description: User not found
  /api/v1/users/{id}/transactions:
    get:
      summary: List transactions for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: RFC3339 timestamp inclusive lower bound
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: RFC3339 timestamp inclusive upper bound
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [income, expense]
          description: Filter operations by type
        - name: category_id
          in: query
          required: false
          schema:
            type: string
          description: Return operations for a specific category
        - name: account_id
          in: query
          required: false
          schema:
            type: string
          description: Return operations for a specific account
      responses:
        '200':
          description: Transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
        '404':
          description: Not found
  /api/v1/users/{id}/reports/overview:
    get:
      summary: Get aggregated spending, income and account balances for a family
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: RFC3339 timestamp inclusive lower bound
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: RFC3339 timestamp inclusive upper bound
      responses:
        '200':
          description: Aggregated report for the selected period
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsOverviewResponse'
        '404':
          description: Not found
  /api/v1/users/{id}/planned-operations:
    get:
      summary: List planned operations for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Planned operations grouped by status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlannedOperationsListResponse'
        '404':
          description: Not found
    post:
      summary: Create a planned operation for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlannedOperationRequest'
      responses:
        '201':
          description: Created planned operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlannedOperationResponse'
        '400':
          description: Validation error
        '404':
          description: Not found
  /api/v1/users/{id}/planned-operations/{operationId}/complete:
    post:
      summary: Mark a planned operation as completed and record the transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: operationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlannedOperationCompleteRequest'
      responses:
        '200':
          description: Planned operation completion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlannedOperationCompleteResponse'
        '400':
          description: Validation error
        '404':
          description: Not found
  /api/v1/transactions:
    post:
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid request
components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name, currency]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        locale:
          type: string
        currency:
          type: string
        family_name:
          type: string
    RegisterResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        family:
          $ref: '#/components/schemas/Family'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
    User:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        locale:
          type: string
        currency_default:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Family:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        currency_base:
          type: string
        created_at:
          type: string
          format: date-time
    Category:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        parent_id:
          type: string
          nullable: true
        name:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        color:
          type: string
        description:
          type: string
        is_archived:
          type: boolean
        is_system:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Account:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [cash, card, bank, deposit, wallet]
        currency:
          type: string
        balance_minor:
          type: integer
          format: int64
        is_archived:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AccountRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cash, card, bank, deposit, wallet]
        currency:
          type: string
        initial_balance_minor:
          type: integer
          format: int64
      required:
        - name
        - type
    AccountResponse:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/Account'
    CategoryRequest:
      type: object
      required: [name, type, color]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        color:
          type: string
        description:
          type: string
        parent_id:
          type: string
          nullable: true
    CategoryResponse:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/Category'
    CategoryArchiveRequest:
      type: object
      required: [archived]
      properties:
        archived:
          type: boolean
    Transaction:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        user_id:
          type: string
        account_id:
          type: string
        category_id:
          type: string
        type:
          type: string
          enum: [income, expense]
        amount_minor:
          type: integer
        currency:
          type: string
        comment:
          type: string
          nullable: true
        occurred_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TransactionRequest:
      type: object
      required: [user_id, account_id, category_id, type, amount_minor, currency, occurred_at]
      properties:
        user_id:
          type: string
        account_id:
          type: string
        category_id:
          type: string
        type:
          type: string
          enum: [income, expense]
        amount_minor:
          type: integer
        currency:
          type: string
        comment:
          type: string
          nullable: true
        occurred_at:
          type: string
          format: date-time
    PlannedOperation:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        user_id:
          type: string
        account_id:
          type: string
        category_id:
          type: string
        type:
          type: string
          enum: [income, expense]
        title:
          type: string
        amount_minor:
          type: integer
        currency:
          type: string
        comment:
          type: string
          nullable: true
        due_at:
          type: string
          format: date-time
        recurrence:
          type: string
          nullable: true
        is_completed:
          type: boolean
        last_completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        creator:
          $ref: '#/components/schemas/FamilyMember'
    PlannedOperationRequest:
      type: object
      required: [account_id, category_id, type, title, amount_minor, due_at]
      properties:
        account_id:
          type: string
        category_id:
          type: string
        type:
          type: string
          enum: [income, expense]
        title:
          type: string
        amount_minor:
          type: integer
        currency:
          type: string
        comment:
          type: string
        due_at:
          type: string
          format: date-time
        recurrence:
          type: string
    PlannedOperationResponse:
      type: object
      properties:
        planned_operation:
          $ref: '#/components/schemas/PlannedOperation'
    PlannedOperationsListResponse:
      type: object
      properties:
        planned_operations:
          type: array
          items:
            $ref: '#/components/schemas/PlannedOperation'
        completed_operations:
          type: array
          items:
            $ref: '#/components/schemas/PlannedOperation'
    PlannedOperationCompleteRequest:
      type: object
      properties:
        occurred_at:
          type: string
          format: date-time
    PlannedOperationCompleteResponse:
      type: object
      properties:
        planned_operation:
          $ref: '#/components/schemas/PlannedOperation'
        transaction:
          $ref: '#/components/schemas/Transaction'
    ReportsOverviewResponse:
      type: object
      properties:
        reports:
          $ref: '#/components/schemas/ReportsOverview'
    ReportsOverview:
      type: object
      properties:
        period:
          $ref: '#/components/schemas/ReportPeriod'
        expenses:
          $ref: '#/components/schemas/MovementReport'
        incomes:
          $ref: '#/components/schemas/MovementReport'
        account_balances:
          type: array
          items:
            $ref: '#/components/schemas/AccountBalanceReport'
      required: [period, expenses, incomes, account_balances]
    ReportPeriod:
      type: object
      properties:
        start_date:
          type: string
          format: date-time
          nullable: true
        end_date:
          type: string
          format: date-time
          nullable: true
    MovementReport:
      type: object
      properties:
        totals:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyAmount'
        by_category:
          type: array
          items:
            $ref: '#/components/schemas/CategoryReportItem'
      required: [totals, by_category]
    CurrencyAmount:
      type: object
      properties:
        currency:
          type: string
        amount_minor:
          type: integer
          format: int64
      required: [currency, amount_minor]
    CategoryReportItem:
      type: object
      properties:
        category_id:
          type: string
        category_name:
          type: string
        category_color:
          type: string
        currency:
          type: string
        amount_minor:
          type: integer
          format: int64
      required: [category_id, category_name, category_color, currency, amount_minor]
    AccountBalanceReport:
      type: object
      properties:
        account_id:
          type: string
        account_name:
          type: string
        account_type:
          type: string
          enum: [cash, card, bank, deposit, wallet]
        currency:
          type: string
        balance_minor:
          type: integer
          format: int64
        is_shared:
          type: boolean
        is_archived:
          type: boolean
      required: [account_id, account_name, account_type, currency, balance_minor, is_shared, is_archived]